Create a correct structure of the python project
Allow to make simple contribuition to the project and add more
modules.
