+++ pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "git-analyzer"
+version = "0.1.0"
+description = "A tool to analyze Git commits and generate articles."
+authors = [
+    { name = "Michael Trimarchi", email = "michael@amarulasolutions.com" },
+]
+dependencies = [
+    "GitPython==3.1.43", # Specify the version used in the original script
+]
+readme = "README.md"
+requires-python = ">=3.8"
+license = { file = "LICENSE" } # You might want to add a LICENSE file
+
+[project.scripts]
+git-analyzer = "main:main"
+
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
