+++ src/__init_.py
@@ -1,0 +1,2 @@
+# This file makes the 'src' directory a Python package.
+# You can leave it empty or add package-level initializations.
